Re-organized everything, bug fixes, more to do
