+++ config.h
@@ -1,0 +1,4 @@
+#ifndef config_defined
+#define config_defined
+    void InitApp(void);
+#endif
