+++ pragmas.h
@@ -1,0 +1,54 @@
+#ifndef pragmas_defined
+#define pragmas_defined
+
+    // Setup bits
+    // PIC18F2431 Configuration Bit Settings
+    // 'C' source line config statements
+    // CONFIG1H
+    #pragma config OSC = HSPLL      // Oscillator Selection bits (HS oscillator, PLL enabled (clock frequency = 4 x FOSC1))
+    #pragma config FCMEN = OFF      // Fail-Safe Clock Monitor Enable bit (Fail-Safe Clock Monitor disabled)
+    #pragma config IESO = OFF       // Internal External Oscillator Switchover bit (Internal External Switchover mode Disabled)
+    // CONFIG2L
+    #pragma config PWRTEN = OFF     // Power-up Timer Enable bit (PWRT disabled)
+    #pragma config BOREN = OFF      // Brown-out Reset Enable bits (Brown-out Reset disabled)
+    // BORV = No Setting
+    // CONFIG2H
+    #pragma config WDTEN = OFF      // Watchdog Timer Enable bit (WDT disabled (control is placed on the SWDTEN bit))
+    #pragma config WDPS = 32768     // Watchdog Timer Postscale Select bits (1:32768)
+    #pragma config WINEN = OFF      // Watchdog Timer Window Enable bit (WDT window disabled)
+    // CONFIG3L
+    #pragma config PWMPIN = OFF     // PWM output pins Reset state control (PWM outputs disabled upon Reset (default))
+    #pragma config LPOL = HIGH      // Low-Side Transistors Polarity (PWM0, 2, 4 and 6 are active-high)
+    #pragma config HPOL = HIGH      // High-Side Transistors Polarity (PWM1, 3, 5 and 7 are active-high)
+    #pragma config T1OSCMX = OFF    // Timer1 Oscillator MUX (Standard (legacy) Timer1 oscillator operation)
+    // CONFIG3H
+    #pragma config MCLRE = ON       // MCLR Pin Enable bit (Enabled)
+    // CONFIG4L
+    #pragma config STVREN = ON      // Stack Full/Underflow Reset Enable bit (Stack full/underflow will cause Reset)
+    #pragma config LVP = ON         // Low-Voltage ICSP Enable bit (Low-voltage ICSP enabled)
+    // CONFIG5L
+    #pragma config CP0 = OFF        // Code Protection bit (Block 0 (000200-000FFFh) not code-protected)
+    #pragma config CP1 = OFF        // Code Protection bit (Block 1 (001000-001FFF)  not code-protected)
+    #pragma config CP2 = OFF        // Code Protection bit (Block 2 (002000-002FFFh) not code-protected)
+    #pragma config CP3 = OFF        // Code Protection bit (Block 3 (003000-003FFFh) not code-protected)
+    // CONFIG5H
+    #pragma config CPB = OFF        // Boot Block Code Protection bit (Boot Block (000000-0001FFh) not code-protected)
+    #pragma config CPD = OFF        // Data EEPROM Code Protection bit (Data EEPROM not code-protected)
+    // CONFIG6L
+    #pragma config WRT0 = OFF       // Write Protection bit (Block 0 (000200-000FFFh) not write-protected)
+    #pragma config WRT1 = OFF       // Write Protection bit (Block 1 (001000-001FFF)  not write-protected)
+    #pragma config WRT2 = OFF       // Write Protection bit (Block 2 (002000-002FFFh) not write-protected)
+    #pragma config WRT3 = OFF       // Write Protection bit (Block 3 (003000-003FFFh) not write-protected)
+    // CONFIG6H
+    #pragma config WRTC = OFF       // Configuration Register Write Protection bit (Configuration registers (300000-3000FFh) not write-protected)
+    #pragma config WRTB = OFF       // Boot Block Write Protection bit (Boot Block (000000-0001FFh) not write-protected)
+    #pragma config WRTD = OFF       // Data EEPROM Write Protection bit (Data EEPROM not write-protected)
+    // CONFIG7L
+    #pragma config EBTR0 = OFF      // Table Read Protection bit (Block 0 (000200-000FFFh) not protected from table reads executed in other blocks)
+    #pragma config EBTR1 = OFF      // Table Read Protection bit (Block 1 (001000-001FFF)  not protected from table reads executed in other blocks)
+    #pragma config EBTR2 = OFF      // Table Read Protection bit (Block 2 (002000-002FFFh) not protected from table reads executed in other blocks)
+    #pragma config EBTR3 = OFF      // Table Read Protection bit (Block 3 (003000-003FFFh) not protected from table reads executed in other blocks)
+    // CONFIG7H
+    #pragma config EBTRB = OFF      // Boot Block Table Read Protection bit (Boot Block (000000-0001FFh) not protected from table reads executed in other blocks)
+
+#endif
